Use BigDecimal for width, length, depth

+++ service/src/main/kotlin/fi/espoo/vekkuli/FormComponents.kt
@@ -3,6 +3,7 @@ package fi.espoo.vekkuli
 import fi.espoo.vekkuli.config.MessageUtil
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
+import java.math.BigDecimal
 
 @Component
 class FormComponents {
@@ -79,7 +80,7 @@ class FormComponents {
     fun decimalInput(
         labelKey: String,
         id: String,
-        value: Double?,
+        value: BigDecimal?,
         required: Boolean? = false,
         attributes: String = "",
         step: Double? = 0.01,
